+++ src/commonMain/kotlin/com/github/blackmo/grass/pot/Plant.kt
@@ -1,0 +1,28 @@
+package com.github.blackmo.grass.pot
+
+import com.github.blackmo.grass.context.GrassParserContext
+import com.github.blackmo.grass.vein.DateTimeTypes
+import kotlinx.coroutines.flow.Flow
+import kotlin.reflect.KClass
+
+/**
+ * Top level implementation of conversion engine of csv contents to **data class**
+ * @param ctx context configuration [GrassParserContext]
+ * @param type data class definition
+ * @author blackmo18
+ */
+@ExperimentalStdlibApi
+ expect  class Plant<T> actual constructor(ctx: com.github.blackmo.grass.context.GrassParserContext, type: KClass<*>):
+    Stem<T> {
+    val dateTimeTypes : DateTimeTypes
+   /**
+    * @return [List&lt;T&gt;] where T is the target data class
+    */
+   fun harvest(seed: List<Map<String, String>>): List<T>
+   /**
+    * @return [Sequence&lt;T&gt;] where T is the target data class
+    */
+   fun harvest(seed: Sequence<Map<String, String>>): Sequence<T>
+
+   suspend fun harvest(seed: Flow<Map<String, String>>): Flow<T>
+}
