+++ src/jvmTest/kotlin/com/github/blackmo/grass/data/DateAndTime.kt
@@ -1,0 +1,11 @@
+package com.github.blackmo.grass.data
+
+import java.time.LocalDate
+import java.time.LocalDateTime
+import java.time.LocalTime
+
+data class DateAndTime(
+    val date: LocalDate,
+    val datetime: LocalDateTime,
+    val time: LocalTime
+)
