+++ src/jvmTest/kotlin/com/github/blackmo/grass/data/DateTime.kt
@@ -1,0 +1,9 @@
+package com.github.blackmo.grass.data
+
+import java.time.LocalDateTime
+import java.time.LocalTime
+
+data class DateTime(
+    val datetime: LocalDateTime,
+    val time: LocalTime
+)
