+++ src/jvmTest/kotlin/com/github/blackmo/grass/data/MissMatch.kt
@@ -1,0 +1,6 @@
+package com.github.blackmo.grass.data
+
+data class MissMatch(
+        val short: Short,
+        val int: Int
+)
