+++ src/jvmTest/kotlin/com/github/blackmo/grass/data/NullableDataTypes.kt
@@ -1,0 +1,11 @@
+package com.github.blackmo.grass.data
+
+data class NullableDataTypes(
+        val short: Short?,
+        val int: Int?,
+        val long: Long?,
+        val float: Float?,
+        val double: Double?,
+        val boolean: Boolean?,
+        val string: String?
+)
