+++ src/jvmTest/kotlin/com/github/blackmo/grass/data/WhiteSpacedStrings.kt
@@ -1,0 +1,11 @@
+package com.github.blackmo.grass.data
+
+data class WhiteSpacedStrings(
+    val short: String,
+    val int: String,
+    val long: String,
+    val float: String,
+    val double: String,
+    val boolean: String,
+    val string: String
+)
