--- src/jvmTest/kotlin/com/vhl/blackmo/grass/data/DateTime.kt
@@ -1,9 +1,0 @@
-package com.vhl.blackmo.grass.data
-
-import java.time.LocalDateTime
-import java.time.LocalTime
-
-data class DateTime(
-    val datetime: LocalDateTime,
-    val time: LocalTime
-)
