--- src/jvmTest/kotlin/com/vhl/blackmo/grass/data/MissMatch.kt
@@ -1,6 +1,0 @@
-package com.vhl.blackmo.grass.data
-
-data class MissMatch(
-        val short: Short,
-        val int: Int
-)
