--- src/jvmTest/kotlin/com/vhl/blackmo/grass/data/NullableDataTypes.kt
@@ -1,11 +1,0 @@
-package com.vhl.blackmo.grass.data
-
-data class NullableDataTypes(
-        val short: Short?,
-        val int: Int?,
-        val long: Long?,
-        val float: Float?,
-        val double: Double?,
-        val boolean: Boolean?,
-        val string: String?
-)
