--- src/jvmTest/kotlin/com/vhl/blackmo/grass/data/WhiteSpacedStrings.kt
@@ -1,11 +1,0 @@
-package com.vhl.blackmo.grass.data
-
-data class WhiteSpacedStrings(
-    val short: String,
-    val int: String,
-    val long: String,
-    val float: String,
-    val double: String,
-    val boolean: String,
-    val string: String
-)
